Work around the broken clisp
The clisp version 1:2.43-1 is broken. It generates a C file with
undefined data types and calls the linker with wrong and unnecessary
options.

The first part copies the contents of the link set “full” to a new
directory and modifies the makevars file. It removes the -lSM and -lICE
flags, because these libraries aren't used in the binary. dpkg-shlibdeps
warned about it. It replaces also the -ldb by -ldb-4.6.

The second part is copied from rte/ordrules/Makefile. The sed call after
the make call was inserted to remove the lines with the unknown data
types. The clisp-link call used the link set with the modified makevars.