dpatch for clisp problems with undefined data types
I found a better way to work around the problems with CLISP. It's
possible to add code to the generated and I use this to include the
header files for the unknown data types and the missing definitions.

The separated make call is still necessary to use the modified linkkit
directory.