A README.source to describe the patch managment system
The new policy version 3.8.0 recommends to include a file that describes
how to use the patch management system of the package. It's allowed to
refer to a file that is installed when all build dependencies are
satisfied. The dpatch package provides such a file since version 2.0.30.

The text was picked from
http://lists.debian.org/debian-policy/2008/08/msg00035.html