Remove and update patches
The file user-commands/texindy was renamed to texindy.in in the last
release and the Makefile.in was modified.

The manual page user-commands/xindy.v2.1 was removed, hence this fix can
go away and because the manual pages of xindy and texindy have been
updated, the problem with hyphens and dashs is gone.