Use dependency on fixed clisp version for xindy package
The memory dump of clisp used by Xindy is not usable across different
version; <http://clisp.cons.org/impnotes/image.html#image-portability>.
Hence, we must ensure the same version used to build the memory dump is
installed at runtime. The package dh-lisp provides a nice script dh_lisp,
but this can't handle memory files;
<http://pkg-common-lisp.alioth.debian.org/clid/clid.html/ch-dh-lisp.html>.
Therefore, I've taken the idea from the script and reimplemented it in my
build scripts.

Every clisp package provides a special virtual package that ensures
exactly this version of clisp is installed. The package xindy must depend
on this version to ensure the clisp version installed at runtime is the
same as the one used to build the memory dump.

The tweaking with the substitution variable clisp:Depends is necessary,
because I found no better way to modify the field misc:Depends in the
file substvars (which is done by the dh_lisp script).